[x11] fix: get desktop bounds correctly (aseprite/aseprite#3118)
Use Xrandr's XRRGetMonitors function to find find the position,
bounds, and "main screen" status of all monitors.

On X11, there is no supported method to find the workarea of a monitor.
This includes the _NET_WORKAREA atom, which breaks multi-monitor
setups. This may not be an issue as most window managers aggressively
and correctly enforce their, albeit private, workarea for windows.